+++ config-model-generator/build-config-models.sh
@@ -12,13 +12,15 @@ then
     # Always delete existing files so when kafka-versions changes we remove
     # models for unsupported versions
     rm ../cluster-operator/src/main/resources/kafka-*-config-model.json || true
+    rm ../topic-operator/src/main/resources/kafka-*-topic-config-model.json || true
 
     for version in "${versions[@]}"
     do
         mvn ${MVN_ARGS} verify exec:java \
         "-Pgenerate-model" \
         "-Dkafka-metadata-version=$version" \
-        "-Dconfig-model-file=../cluster-operator/src/main/resources/kafka-${version}-config-model.json"
+        "-Dconfig-model-file=../cluster-operator/src/main/resources/kafka-${version}-config-model.json" \
+        "-Dtopic-config-model-file=../topic-operator/src/main/resources/kafka-${version}-topic-config-model.json"
     done
 else
     # Clean up the last version in the file?