+++ systemtest/src/test/java/io/strimzi/systemtest/operators/CustomResourceStatusIsolatedST.java
@@ -645,6 +645,7 @@ class CustomResourceStatusIsolatedST extends AbstractST {
         assertThat(kafkaTopicStatus.getConditions().stream()
             .anyMatch(condition -> condition.getReason().equals("InvalidRequestException")), is(true));
         assertThat(kafkaTopicStatus.getConditions().stream()
-            .anyMatch(condition -> condition.getMessage().contains(String.format("Invalid value %s for configuration min.insync.replicas", invalidValue))), is(true));
+            .anyMatch(condition -> condition.getMessage().contains(String.format("KafkaTopic %s/%s has invalid spec.config: " +
+                    "min.insync.replicas has value '%s' which is not an int", Constants.INFRA_NAMESPACE, topicName, invalidValue))), is(true));
     }
 }